Moved metaroom_xml_parser headers into include/metaroom_xml_parser and adapted other projects
