+++ calibrate_sweeps/src/calibrate_sweep_action_server.cpp
@@ -13,7 +13,7 @@
 #include <semantic_map/room_utilities.h>
 #include <semantic_map/reg_transforms.h>
 #include <semantic_map/sweep_parameters.h>
-#include "load_utilities.h"
+#include <metaroom_xml_parser/load_utilities.h>
 #include <strands_sweep_registration/RobotContainer.h>
 
 typedef pcl::PointXYZRGB PointType;
