+++ metaroom_xml_parser/package.xml
@@ -17,6 +17,7 @@
   <build_depend>tf</build_depend>
   <build_depend>std_msgs</build_depend>
   <build_depend>sensor_msgs</build_depend>
+  <build_depend>pcl_ros</build_depend>
 
   <run_depend>rospy</run_depend>
   <run_depend>roscpp</run_depend>
@@ -28,6 +29,7 @@
   <run_depend>tf</run_depend>
   <run_depend>std_msgs</run_depend>
   <run_depend>sensor_msgs</run_depend>
+  <run_depend>pcl_ros</run_depend>
 
   <export>
