+++ metaroom_xml_parser/src/load_labelled_data.cpp
@@ -1,4 +1,4 @@
-#include "load_utilities.h"
+#include "metaroom_xml_parser/load_utilities.h"
 #include <pcl/visualization/pcl_visualizer.h>
 #include <pcl_ros/transforms.h>
