+++ metaroom_xml_parser/src/load_single_file_main.cpp
@@ -1,4 +1,4 @@
-#include "simple_xml_parser.h"
+#include "metaroom_xml_parser/simple_xml_parser.h"
 
 typedef pcl::PointXYZRGB PointType;
