+++ metaroom_xml_parser/src/simple_xml_parser.cpp
@@ -1,3 +1,3 @@
-#include "simple_xml_parser.h"
+#include "metaroom_xml_parser/simple_xml_parser.h"
 
 template class SimpleXMLParser<pcl::PointXYZRGB>;
